Create starter
>>>Hi
This @AGI_communication_bot

provides links to the SingularityNET network. 

It's in development and want to be treaten as a fast in group access and information provider. 

Especially for experienced users. 
It reply to keywords. 

For example:
Website = www.singularitynet.io 

I think it's a good idea to collect trigger words here. And discuss their need.

To set it up just add it to your group and give it Admins status. 


Thanks for collaboration.
Feel free to DM me for further questions or ideas.

https://t.me/joinchat/Gops-U-HT3IIoBrjjNwNSg